refactor button imports and removed rendering files
