+++ packages/react-cap-theme/src/components/react-button/Button.ts
@@ -1,7 +1,3 @@
-export { buttonClassNames } from '@fluentui/react-button';
-export { Button } from './components/Button/Button';
-export { renderButton } from './components/Button/renderButton';
-export { useButton } from './components/Button/useButton';
 export { useButtonStyles } from './components/Button/useButtonStyles.styles';
 export type {
   ButtonProps,
