+++ packages/react-cap-theme/src/components/react-button/MenuButton.ts
@@ -1,7 +1,3 @@
-export { menuButtonClassNames } from '@fluentui/react-button';
-export { MenuButton } from './components/MenuButton/MenuButton';
-export { renderMenuButton } from './components/MenuButton/renderMenuButton';
-export { useMenuButton } from './components/MenuButton/useMenuButton';
 export { useMenuButtonStyles } from './components/MenuButton/useMenuButtonStyles.styles';
 export type {
   MenuButtonProps,
