+++ packages/react-cap-theme/src/components/react-button/SplitButton.ts
@@ -1,7 +1,3 @@
-export { splitButtonClassNames } from '@fluentui/react-button';
-export { SplitButton } from './components/SplitButton/SplitButton';
-export { renderSplitButton } from './components/SplitButton/renderSplitButton';
-export { useSplitButton } from './components/SplitButton/useSplitButton';
 export { useSplitButtonStyles } from './components/SplitButton/useSplitButtonStyles.styles';
 export type {
   SplitButtonProps,
