+++ packages/react-cap-theme/src/components/react-button/ToggleButton.ts
@@ -1,7 +1,3 @@
-export { toggleButtonClassNames } from '@fluentui/react-button';
-export { ToggleButton } from './components/ToggleButton/ToggleButton';
-export { renderToggleButton } from './components/ToggleButton/renderToggleButton';
-export { useToggleButton } from './components/ToggleButton/useToggleButton';
 export { useToggleButtonStyles } from './components/ToggleButton/useToggleButtonStyles.styles';
 export type {
   ToggleButtonProps,
