--- packages/react-cap-theme/src/components/react-button/components/ToggleButton/renderToggleButton.tsx
@@ -1,6 +1,0 @@
-import type { JSXElement } from '@fluentui/react-utilities';
-import { renderButton } from '../../Button';
-import type { ToggleButtonState } from './ToggleButton.types';
-
-export const renderToggleButton = (state: ToggleButtonState): JSXElement =>
-  renderButton(state);
