+++ packages/react-cap-theme/src/components/react-button/index.ts
@@ -1,4 +1,24 @@
-export { useButtonStyles } from './components/Button/useButtonStyles.styles';
-export { useMenuButtonStyles } from './components/MenuButton/useMenuButtonStyles.styles';
-export { useSplitButtonStyles } from './components/SplitButton/useSplitButtonStyles.styles';
-export { useToggleButtonStyles } from './components/ToggleButton/useToggleButtonStyles.styles';
+export { useButtonStyles } from './Button';
+export type {
+  ButtonProps,
+  ButtonSlots,
+  ButtonState,
+  ButtonAppearance,
+} from './Button';
+
+export { useMenuButtonStyles } from './MenuButton';
+export type {
+  MenuButtonProps,
+  MenuButtonSlots,
+  MenuButtonState,
+} from './MenuButton';
+
+export { useSplitButtonStyles } from './SplitButton';
+export type {
+  SplitButtonProps,
+  SplitButtonSlots,
+  SplitButtonState,
+} from './SplitButton';
+
+export { useToggleButtonStyles } from './ToggleButton';
+export type { ToggleButtonProps, ToggleButtonState } from './ToggleButton';
